# packages/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Nexus API
  description: Local-first AI document analysis and knowledge graph platform
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: local development environment
paths:
...
  /workspaces:
    get:
      summary: List all workspaces
      tags: [workspaces]
      operationId: listWorkspaces
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'

    post:
      summary: Create a new workspace
      tags: [workspaces]
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workspaces/{workspaceId}:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get workspace by ID
      tags: [workspaces]
      operationId: getWorkspace
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update workspace
      tags: [workspaces]
      operationId: updateWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete workspace
      tags: [workspaces]
      operationId: deleteWorkspace
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'
...
  /workspaces/{workspaceId}/chats/{chatId}:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: chatId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get chat details with message history
      operationId: getChat
      tags: [chats]
      parameters:
        - name: message_limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          description: Number of recent messages to return
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [chat, messages]
                properties:
                  chat:
                    $ref: '#/components/schemas/Chat'
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a chat (soft delete)
      operationId: deleteChat
      tags: [chats]
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'
...
  /workspaces/{workspaceId}/chats/{chatId}/sources:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: chatId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    
    get:
      summary: Get sources used in chat messages
      operationId: getChatSources
      tags: [chats]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourcesResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ========================================
  # Graph CRUD Operations
  # ========================================
...
components:
  schemas:
    Workspace:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        settings:
          type: object
          nullable: true
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Chat:
      type: object
      required: [id, workspace_id, title, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        title:
          type: string
        filter_config:
          $ref: '#/components/schemas/FilterConfig'
        message_count:
          type: integer
          minimum: 0
          description: Total number of messages in this chat
        last_message_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the last message
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChatMessage:
      type: object
      required: [id, chat_id, role, content, message_index, created_at]
      properties:
        id:
          type: string
          format: uuid
        chat_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        message_index:
          type: integer
          minimum: 0
          description: Sequential index within the chat (0-based)
        document_refs:
          type: array
          items:
            $ref: '#/components/schemas/DocumentReference'
          description: References to source documents (RAG citations)
        created_at:
          type: string
          format: date-time
    DocumentReference:
      type: object
      required: [document_id, chunk_index, score]
      properties:
        document_id:
          type: string
          format: uuid
        document_name:
          type: string
        chunk_index:
          type: integer
          minimum: 0
        # ページ番号を追加：フロントのソースパネルで「P.3」表示に使う
        page_number:
          type: integer
          minimum: 1
          nullable: true
          description: Page number in the source document (1-based)
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Relevance score (0.0 - 1.0)
        # content_preview は廃止せず残すが、上限を緩和しない
        # 全文表示は GET /documents/{id}/pages/{page} で別途取得する設計（関心の分離）
        content_preview:
          type: string
          maxLength: 200
          description: Short preview of the chunk content
...
    SourceDocument:
      type: object
      required:
        - document_id
        - document_name
        - mime_type
        - chunks
      properties:
        document_id:
          type: string
          format: uuid
        document_name:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
          format: int64
        # ページ単位のグルーピング用：このドキュメント内で参照されたページ番号一覧
        referenced_pages:
          type: array
          items:
            type: integer
          description: Page numbers referenced in this document (deduplicated, sorted)
        chunks:
          type: array
          items:
            type: object
            required: [chunk_index, score, content_preview]
            properties:
              chunk_index:
                type: integer
              page_number:
                type: integer
                minimum: 1
                nullable: true
                description: Page number this chunk belongs to (1-based)
              score:
                type: number
                format: float
              content_preview:
                type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SourcesResponse:
      type: object
      required:
        - sources
        - total_documents
        - total_chunks
      properties:
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceDocument'
        total_documents:
          type: integer
        total_chunks:
          type: integer
    FilterConfig:
      type: object
      properties:
        document_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Specific documents to search (empty = all)
        directory_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Search within specific directories
        tags:
          type: array
          items:
            type: string
          description: Filter by document tags
      description: Configuration for filtering RAG search scope

    # ========================================
    # Analysis Schemas
    # ========================================
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: 
              type: string
            message: 
              type: string
            details: 
              type: object  
            timestamp: 
              type: string
            request_id: 
              type: string  
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...